Move the pane's view into App/Terminal
App/ had every control flat at the root, and this branch added three
more. Put the pane's view where its structure is visible in the tree:

  App/Terminal/            TerminalControl (composite), SurfaceHost
  App/Terminal/Overlays/   PaneStatusOverlay, ScrollbackOverlay, SearchOverlay

Tabs/Panes/ stays the pane *tree* (model + SplitPanel); Terminal/ is
one pane's *view*. Includes take the folder path; the generated .g.h
/ .g.cpp includes accept either the mirrored subfolder or the root
until the first build settles which one the XAML compiler emits.
