+++ App/Terminal/Overlays/PaneStatusOverlay.idl
