+++ App/Terminal/Overlays/PaneStatusOverlay.xaml
