+++ App/Terminal/Overlays/ScrollbackOverlay.xaml
