+++ App/Terminal/TerminalControl.xaml
