+++ App/Terminal/TerminalControl.xaml.cpp
@@ -1,8 +1,10 @@
 #include "pch.h"
-#include "TerminalControl.xaml.h"
+#include "Terminal/TerminalControl.xaml.h"
 #include "resource.h"
 #include "Interop/Encoding.h"
-#if __has_include("TerminalControl.g.cpp")
+#if __has_include("Terminal/TerminalControl.g.cpp")
+#include "Terminal/TerminalControl.g.cpp"
+#elif __has_include("TerminalControl.g.cpp")
 #include "TerminalControl.g.cpp"
 #endif
